directory restructure, initial css copied over, started fixing css.
