--- client/src/App/index.js
@@ -1,1 +1,0 @@
-export * from './App';
