--- client/src/_webpages/HomePage/index.js
@@ -1,1 +1,0 @@
-export * from './HomePage';
