--- client/src/_webpages/LoginPage/index.js
@@ -1,1 +1,0 @@
-export * from './LoginPage';
