--- client/src/_webpages/SignupPage/index.js
@@ -1,1 +1,0 @@
-export * from './SignupPage';
