+++ client/src/components/PasswordValidator.js
