+++ client/src/index.html
@@ -3,17 +3,17 @@
 <head>
     <meta charset="UTF-8">
     <title>Movie Together</title>
-    <link rel="stylesheet" href="https://netdna.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" />
+    <!-- <link rel="stylesheet" href="https://netdna.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" />
     <style>
         a { cursor: pointer; }
         .help-block { font-size: 12px; }
-    </style>
+    </style> -->
 </head>
 <body>
     <div id="app"></div>
 
     <!-- credits -->
-    <div class="text-center">
+    <div className="github_link">
         <p>
             <a href="https://github.com/c-salt/movie-website" target="_top">Movie Together Website Github</a>
         </p>
