+++ client/src/index.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import { App } from './App';
+import { App } from './components/App';
+import './style/style.less'
 
 render(
     <App />,
