+++ client/src/style/components_style/helpbox.less
@@ -1,0 +1,150 @@
+.password_validator {
+  .topLeft(0, 105%);
+  position: absolute;
+  display: block;
+  opacity: 0;
+  .transition(all 0.2s ease-in-out);
+  .size(320px, auto);
+  .box-sizing(border-box);
+  .user-select(none);
+  .transition(all 0.2s ease-in-out);
+  opacity: 1;
+  z-index: @layer3;
+
+
+  &:after {
+    .display(absolute);
+    .topLeft(30px, -10px);
+    content: '';
+    border-top: 10px solid transparent;
+    border-bottom: 10px solid transparent;
+    border-right: 10px solid #e4e7e8;
+  }
+
+  .validator_container {
+    .display(relative);
+    .box-sizing(border-box);
+    .box-sizing(border-box);
+    .border-radius(4px);
+    background: #e4e7e8;
+    padding: 10px 0px 10px 20px;
+
+    h4.validator_title {
+      .display(relative);
+      .transition(all 0.2s ease-in-out);
+      margin: 12px 0 5px 0;
+      font-size: 12px;
+      font-weight: 500;
+      text-transform: uppercase;
+      color: @dark-blue;
+      opacity: .25;
+      letter-spacing: 1.5px;
+
+      &.valid {
+        color: @green;
+        opacity: 1
+      }
+    }
+
+    ul.rules_list {
+      .display(relative);
+      margin: 0;
+      padding: 0;
+
+      > li {
+        .display(relative);
+        margin: 12px 0 12px 0;
+        font-size: 12px;
+        letter-spacing: .3px;
+        font-weight: 200;
+        color: @dark-blue;
+
+        .bad_word {
+          font-style: italic;
+          padding: 0 5px 0 0;
+
+          &:after {
+            content: ',';
+            display: inline;
+          }
+
+          &:last-child {
+            padding: 0;
+
+            &:after {
+              display: none;
+            }
+          }
+        }
+
+        .icon_invalid {
+          .display(absolute);
+          .topLeft(1px, 0);
+          .scale(1);
+          opacity: 1;
+          .size(18px, 18px);
+          .transition(all 0.2s ease-in-out);
+
+          svg {
+            opacity: 1;
+            .size(18px, 18px);
+
+            path {
+              fill: @red;
+            }
+          }
+        }
+
+        .icon_valid {
+          position: absolute;
+          .topLeft(1px, 0);
+          .scale(0);
+          opacity: 0;
+          .size(19px, 19px);
+          .transition(all 0.2s ease-in-out);
+
+          svg {
+            opacity: 1;
+            .size(19px, 19px);
+          }
+        }
+
+        .error_message {
+          .display(relative);
+          .transition(all 0.2s ease-in-out);
+          padding: 0 0 0 27px;
+          opacity: .9;
+        }
+
+        &.valid {
+          .icon_invalid {
+            .scale(0);
+            opacity: 0;
+          }
+
+          .icon_valid {
+            .scale(1);
+            opacity: 1;
+          }
+          
+          .error_message {
+            opacity: .3;
+          }
+
+        }
+      }
+    }
+  }
+
+  &.visible {
+    .transform(translateX(0px));
+    opacity: 1;
+  }
+
+  &.invisible {
+    .transform(translateX(-15px));
+    opacity: 0;
+    -webkit-animation-delay: 2s; /* Chrome, Safari, Opera */
+    animation-delay: 2s;
+  }
+}
