+++ client/src/style/components_style/scrollbar.less
@@ -1,0 +1,67 @@
+.webkit-scrollbar-light() {
+  &::-webkit-scrollbar {
+    width: 16px;
+    background-color: transparent;
+    background-clip: content-box;
+    cursor: pointer;
+  }
+
+  &::-webkit-scrollbar-button {
+    background-color: transparent;
+  }
+  &::-webkit-scrollbar-corner {
+    background-color: transparent;
+  }
+  &::-webkit-scrollbar-thumb {
+    border: solid transparent;
+    border-width: 6px;
+    background-color: #c5c5c5;
+    background-clip: content-box;
+    cursor: pointer;
+
+    &:hover {
+      background-color: darken(#ccc, 5%);
+    }
+  }
+  &::-webkit-scrollbar-track {
+    border: solid #fff;
+    border-width: 1px 1px 1px 2px;
+    background-color: #fff;
+    background-clip: content-box;
+  }
+  &::-webkit-scrollbar-track-piece {
+    border-left: 1px solid @gray;
+    background-color: transparent;
+  }
+
+  &::-webkit-scrollbar:horizontal {
+    height: 12px;
+    background-color: transparent;
+    background-clip: content-box;
+  }
+
+  &::-webkit-scrollbar-track:horizontal {
+    background-color: #fff;
+    background-clip: content-box;
+  }
+  &::-webkit-scrollbar-track-piece:horizontal {
+    border-top: 1px solid transparent;
+    border-left: 1px solid transparent;
+  }
+  &::-webkit-scrollbar-thumb:horizontal {
+    border: solid transparent;
+    border-width: 2px 1px 2px 2px;
+    background-color: #c5c5c5;
+    background-clip: content-box;
+
+    &:hover {
+      background-color: darken(#ccc, 5%);
+    }
+  }
+  &::-webkit-scrollbar-corner {
+    display: none
+  }
+  &::-webkit-resizer {
+    display: none;
+  }
+}
