+++ client/src/style/screens_style/app.less
@@ -1,0 +1,3 @@
+.container {
+  background-color: orange
+}
