+++ client/src/style/screens_style/signup_page.less
@@ -1,0 +1,83 @@
+.create_account_screen {
+  .display(absolute);
+  .topLeft(0, 0);
+  .bottomRight(0, 0);  
+  min-height: 520px;
+  opacity: 0;
+  .animation(fadeIn .4s ease-in-out forwards);
+
+  .create_account_form {
+    .flex-display();
+    .flex-direction(column);
+    .justify-content(center);
+    .align-items(center);
+    position: relative;
+    margin: 80px 0 200px 0;
+    min-height: 400px;
+
+    > h1 {
+      .display(relative);
+      font-size: 34px;
+      letter-spacing: .8px;
+      font-weight: 200;
+      line-height: 34px;
+      margin: 0 0 5px 0;
+    }
+
+    p {
+      .display(relative);
+      font-size: 16px;
+      font-weight: 100;
+      letter-spacing: .7px;
+      color: @dark-blue;
+      text-align: center;
+      opacity: .5;
+      margin: 0 0 30px 0;
+    }
+
+    .checkbox_group {
+      .display(relative);
+      min-height: 40px;
+
+      span {
+        letter-spacing: .5px;
+      }
+    }
+
+    .github_link {
+      .display(relative);
+      margin: 50px 0 50px 0;
+      opacity: 0;
+      .animation(fadeIn .4s .3s ease-in-out forwards);
+
+      svg {
+        .size(26px, 26px);
+
+        path {
+          .transition(all 0.2s ease-in-out);
+          fill: #545a6a;
+        }
+      }
+
+      &:hover {
+        svg {
+          path {
+            fill: @light-blue;
+          }
+        }
+      }
+    }
+  }
+
+  form {
+    width: 430px;
+
+    @media @mobile {
+      width: 300px;
+    }
+  }
+}
+
+@-webkit-keyframes fadeIn { from { opacity:0; } to { opacity:1; } }
+@-moz-keyframes fadeIn { from { opacity:0; } to { opacity:1; } }
+@keyframes fadeIn { from { opacity:0; } to { opacity:1; } }
