+++ client/src/style/style.less
@@ -1,0 +1,22 @@
+// UI Setup - Vars, Resets, Fonts, Mixins
+// ======================================================
+
+
+@import "ui_setup.less";
+
+
+// UI Components - Reusable elements
+//========================================================
+
+
+@import "ui_components.less";
+
+
+// App - main styles file
+//=======================================================
+
+
+@import "app_scaffolding.less";
+
+
+// =======================================================
