+++ client/src/style/ui_components.less
@@ -1,0 +1,7 @@
+@import "components_style/basics.less";
+@import "components_style/input.less";
+@import "components_style/input_error.less";
+@import "components_style/helpbox.less";
+@import "components_style/select.less";
+@import "components_style/scrollbar.less";
+@import "components_style/button.less";
