+++ client/src/style/ui_config/colors.less
@@ -1,0 +1,55 @@
+// Main colors
+// -------------------------
+@light-blue:                            #40b4de;
+@dark-blue:                             #363b4a;
+@green:                                 #50b87f;
+@gray:                                  #f0f4f5;
+@black:                                 #242424;
+@orange:                                #f16852;
+@red:                                   #F16767;
+@yellow:                                #ffd780;
+@disabled:                              #c4c4c4;
+@white:                                 #ffffff;
+
+
+// Backgrounds
+// -------------------------
+@body-bg:                               @gray;
+@gray-bg:                               #E3E5E6;
+
+
+// Links
+// -------------------------
+@link-color:                            @light-blue;
+@link-hover-color:                      darken(@light-blue,20%);
+@link-active-color:                     lighten(@light-blue,20%);
+
+
+// Texts
+// -------------------------
+@text-dark-blue:                        #494f5c;
+@text-gray-blue:                        #acb2c0;
+@sidebar-text-color:                    #a8aab3;
+
+
+// Elements
+// -------------------------
+@icon-gray:                             #acb2c0;
+@stroke-gray:                           #d4d8d9;
+
+
+// Buttons
+// -------------------------
+@button-default-color:                  #fff;
+@button-default-bg:                     #545a6a;
+@button-switch-inactive:                #86909e;
+@button-hover-bg:                       lighten(@dark-blue,10%);
+@button-active-bg:                      @light-blue;
+@button-disabled:                       #c4c4c4;
+@button-gray:                           #dadede;
+@button-lite-color:                     #a6a6a6;
+@fue-button-color:                      #67aadd;
+
+
+
+
