+++ client/src/style/ui_config/media_queries.less
@@ -1,0 +1,19 @@
+// Media Queries Breakpoints
+// -------------------------
+
+//@screen1024:      ~"only screen and (min-width: 760px) and (max-width: 1024px)"; uncomment after optimization for mobile
+@screen1024:      ~"only screen and (min-width: 871px) and (max-width: 1024px)";
+@screen1280:      ~"only screen and (min-width: 1025px) and (max-width: 1280px)";
+@screen1366:      ~"only screen and (min-width: 1281px) and (max-width: 1366px)";
+@screen1440:      ~"only screen and (min-width: 1367px) and (max-width: 1440px)";
+@screen1680:      ~"only screen and (min-width: 1441px) and (max-width: 1680px)";
+@screen1920:      ~"only screen and (min-width: 1681px) and (max-width: 1920px)";
+@screen2560:      ~"only screen and (min-width: 1921px)";
+@mobile:          ~"only screen and (min-width : 0px) and (max-width : 870px)";
+
+
+// Targeted to devices
+
+@iPhone:          ~"only screen and (-webkit-min-device-pixel-ratio : 1.5), only screen and (min-device-pixel-ratio : 1.5)";
+@iPadLandscape:   ~"only screen and (min-device-width : 768px) and (max-device-width : 1024px) and (orientation : landscape)";
+@iPadPortrait:    ~"only screen and (min-device-width : 768px) and (max-device-width : 1024px) and (orientation : portrait)";
