+++ client/src/style/ui_config/typography.less
@@ -1,0 +1,12 @@
+// Code
+@font-monospace:                 Monaco, Menlo, Consolas, "Courier New", monospace;
+@font-size-base:                 14px;
+@line-height-base:               1.6;
+
+@line-height-computed:           floor(@font-size-base * @line-height-base);
+@font-size-h1:                   floor(@font-size-base * 3.444);
+@font-size-h2:                   ceil(@font-size-base * 2.889);
+@font-size-h3:                   ceil(@font-size-base * 2.222);
+@font-size-h4:                   ceil(@font-size-base * 1.611);
+@font-size-h5:                   floor(@font-size-base * 1.556);
+@font-size-h6:                   ceil(@font-size-base * 1.333);
