+++ client/src/style/ui_config/zindex_layers.less
@@ -1,0 +1,9 @@
+@layer0: -1;
+@layer1: 1;
+@layer2: 10;
+@layer3: 100;
+@layer4: 1000;
+@layer5: 10000;
+
+@layer-bottom: @layer1;
+@layer-top: @layer5;
