+++ client/src/style/ui_setup.less
@@ -1,0 +1,12 @@
+// UI variables
+@import "ui_config/colors.less";
+@import "ui_config/typography.less";
+@import "ui_config/media_queries.less";
+@import "ui_config/zindex_layers.less";
+
+// Reset
+@import "ui_config/reset.less";
+
+// LESS Mixins
+@import "ui_config/mixins.less";
+@import "ui_config/flexMixins.less";
