+++ client/src/webpages/LoginPage/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { userService } from '../../_services';
+import { userService } from '../../services';
 
 class LoginPage extends React.Component {
     constructor(props) {
